requirements.txt
annotated-types==0.7.0
anyio==4.7.0
audioread==3.0.1
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fonttools==4.55.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
matplotlib==3.10.0
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
passlib==1.7.4
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.36
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0